+++ packages/article-label/__tests__/web/jest.config.js
@@ -1,5 +1,9 @@
 const jestConfigurator = require("@times-components/jest-configurator");
 
 module.exports = Object.assign(jestConfigurator("article-label", "web"), {
-  collectCoverageFrom: ["**/packages/article-label/article-label.js"]
+  collectCoverageFrom: [
+    "**/packages/article-label/article-label.js",
+    "**/packages/article-label/style/*.js",
+    "!**/packages/article-label/**/*.android.js"
+  ]
 });