+++ packages/link/__tests__/ios/jest.config.js
@@ -2,10 +2,9 @@ const jestConfigurator = require("@times-components/jest-configurator");
 
 module.exports = Object.assign(jestConfigurator("link", "ios"), {
   collectCoverageFrom: [
-    "**/packages/link/**/*.js",
-    "!**/packages/link/**/link.web.js",
-    "!**/packages/link/**/link.android.js",
-    "!**/packages/link/**/*stories*",
+    "**/packages/link/*.js",
+    "!**/packages/link/*.web.js",
+    "!**/packages/link/*.stories.js",
     "!**/packages/link/__tests__/**",
     "!**/packages/link/coverage/**"
   ]