+++ packages/link/__tests__/web/jest.config.js
@@ -2,11 +2,8 @@ const jestConfigurator = require("@times-components/jest-configurator");
 
 module.exports = Object.assign(jestConfigurator("link", "web"), {
   collectCoverageFrom: [
-    "**/packages/link/**/*.js",
-    "!**/packages/link/**/link.js",
-    "!**/packages/link/**/link.android.js",
-    "!**/packages/link/**/*stories*",
-    "!**/packages/link/__tests__/**",
-    "!**/packages/link/coverage/**"
+    "**/packages/link/*.web.js",
+    "**/packages/link/text-link.js",
+    "!**/packages/link/*.stories.js"
   ]
 });