+++ packages/article/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.24.7"></a>
+## [0.24.7](https://github.com/newsuk/times-components/compare/@times-components/article@0.24.6...@times-components/article@0.24.7) (2018-01-05)
+
+
+
+
+**Note:** Version bump only for package @times-components/article
+
+<a name="0.24.6"></a>
+## [0.24.6](https://github.com/newsuk/times-components/compare/@times-components/article@0.24.5...@times-components/article@0.24.6) (2018-01-05)
+
+
+
+
+**Note:** Version bump only for package @times-components/article
+
+<a name="0.24.5"></a>
+## [0.24.5](https://github.com/newsuk/times-components/compare/@times-components/article@0.24.4...@times-components/article@0.24.5) (2018-01-05)
+
+
+
+
+**Note:** Version bump only for package @times-components/article
+
+<a name="0.24.4"></a>
+## [0.24.4](https://github.com/newsuk/times-components/compare/@times-components/article@0.24.3...@times-components/article@0.24.4) (2018-01-05)
+
+
+### Bug Fixes
+
+* make styling updates to article header and article meta ([#492](https://github.com/newsuk/times-components/issues/492)) ([55537f8](https://github.com/newsuk/times-components/commit/55537f8))
+
+
+
+
+<a name="0.24.3"></a>
+## [0.24.3](https://github.com/newsuk/times-components/compare/@times-components/article@0.24.2...@times-components/article@0.24.3) (2018-01-05)
+
+
+### Bug Fixes
+
+* REPLAT-1167 update web article snapshots ([#497](https://github.com/newsuk/times-components/issues/497)) ([60fea21](https://github.com/newsuk/times-components/commit/60fea21))
+
+
+
+
 <a name="0.24.2"></a>
 ## [0.24.2](https://github.com/newsuk/times-components/compare/@times-components/article@0.24.1...@times-components/article@0.24.2) (2018-01-04)